fix 120-ish build warnings on Windows using gsl::narrow_cast
On Windows we get warned of implicit casts from size_t to int (and some other potentially lossy casts).

Rather than peppering the code with a bunch more static_cast's, using narrow_cast<> from the ISO Guidelines Support Library (GSL), via Microsoft's open-source header-only implementation.

narrow_cast<> better expresses the intent of these casts (namely, we believe these are safe). It also has the benefit of being searchable and distinguishes them from other uses of static_cast.

The GSL also has narrow<>, which checks that the cast is safe and throws if it is not; I didn't use this, but we could consider using it in debug-builds to help confirm our assumptions (not part of this change).

The GSL is related to the ISO's C++ Core Guidelines, see http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines.

The Guideline Support Library puts a few other interesting tidbits at our disposal, see: http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#S-gsl

In my debug Windows build in Qt Creator, takes the overall warning count from 271 down to 142. Most of the remainder are in 3rd party code I don't want to touch, or different warnings that need further investigation.

+++ src/cpp/r/RUtil.cpp
@@ -1,7 +1,7 @@
 /*
  * RUtil.cpp
  *
- * Copyright (C) 2009-18 by RStudio, Inc.
+ * Copyright (C) 2009-19 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -12,7 +12,6 @@
  * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
  *
  */
-
 
 #include <r/RUtil.hpp>
 
@@ -28,6 +27,7 @@
 #include <r/RExec.hpp>
 
 #include <R_ext/Riconv.h>
+#include <gsl/gsl_util>
 
 #ifndef CP_ACP
 # define CP_ACP 0
@@ -182,7 +182,7 @@ core::Error iconvstr(const std::string& value,
             ::Riconv_close(handle);
             Error error = systemError(R_ERRNO, ERROR_LOCATION);
             error.addProperty("str", value);
-            error.addProperty("len", value.length());
+            error.addProperty("len", gsl::narrow_cast<int>(value.length()));
             error.addProperty("from", from);
             error.addProperty("to", to);
             return error;
